+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 ## v2.1.0.0
 - Change uix of related videos carousel
+- Implement nunito font and remove unused font files
 
 ## v2.0.23.15
 - Reduce progress bar, and fix it for IE11